requirements.txt
discord.py --- REWRITE


# ex for voice: python3 -m pip install -U git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py[voice]